create: Add oidc-provider command
STS clusters require an OIDC provider so that operators can authenticate
against the customer's AWS account. This new command uses the generated
OIDC endpoint URL from OCM and creates the provider in the customer's
AWS account. Alternatively it outputs the commands necessary so that the
customer can create them manually.