account-roles: Ensure that roles and policies can be upgraded
If a customer creates roles and policies for OpenShift 4.7 clusters and
then decides to upgrade to 4.8, we need to ensure that they can re-run
the initialization commands that will update the resources to be
compatible. Since later versions are a superset of older versions, we
enforce upgrading and ignore downgrade requests.